requirements.txt
beautifulsoup4==4.11.2
bleach==6.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.2
docutils==0.19
idna==3.4
importlib-metadata==6.0.0
jaraco.classes==3.2.3
jeepney==0.8.0
keyring==23.13.1
lxml==4.9.2
markdown-it-py==2.2.0
mdurl==0.1.2
more-itertools==9.1.0
pkginfo==1.9.6
pycparser==2.21
Pygments==2.14.0
readme-renderer==37.3
requests==2.28.2
requests-toolbelt==0.10.1
rfc3986==2.0.0
rich==13.3.1
SecretStorage==3.3.3
six==1.16.0
soupsieve==2.4
twine==4.0.2
urllib3==1.26.14
webencodings==0.5.1
zipp==3.15.0
selenium==4.8.2
pillow==9.4.0
chromedriver-binary